requirements.txt
django==2
pillow
django-crispy-forms
paywix
requests
django-simple-history
django-extensions
pyparsing
pydot
graphviz
pydot-ng
Werkzeug
